+++ packages/terra-application-layout/src/terra-dev-site/test/application-layout/ApplicationLayoutHeaderWithIcons.test.jsx
@@ -1,14 +1,16 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
+import Icon1 from 'terra-icon/lib/icon/IconTrash';
+import Icon2 from 'terra-icon/lib/icon/IconFolder';
+import Icon3 from 'terra-icon/lib/icon/IconSearch';
 import ApplicationHeader from '../../../header/_ApplicationHeader';
-import ExtensionsExample from '../common/ExtensionsExample';
 
-const ApplicationLayoutHeaderSmall = () => (
+const ApplicationLayoutHeader = () => (
   <MemoryRouter>
     <ApplicationHeader
       id="test-header"
-      layoutConfig={{ size: 'small', toggleMenu: () => {} }}
+      layoutConfig={{ size: 'large' }}
       nameConfig={{ title: 'app-test-title' }}
       utilityConfig={{
         title: 'test-util-title',
@@ -22,23 +24,25 @@ const ApplicationLayoutHeaderSmall = () => (
           },
         ],
       }}
-      extensions={<ExtensionsExample layoutConfig={{ size: 'large' }} />}
       applicationLinks={{
         links: [
           {
             id: '123',
             path: '/something1',
             text: 'item 1',
+            icon: <Icon1 />,
           },
           {
             id: '234',
             path: '/something2',
             text: 'item 2',
+            icon: <Icon2 />,
           },
           {
             id: '345',
             path: '/something3',
             text: 'item 3',
+            icon: <Icon3 />,
           },
         ],
       }}
@@ -46,4 +50,4 @@ const ApplicationLayoutHeaderSmall = () => (
   </MemoryRouter>
 );
 
-export default ApplicationLayoutHeaderSmall;
+export default ApplicationLayoutHeader;