+++ packages/terra-application-links/src/terra-dev-site/doc/example/ApplicationTabsWithIconsExample.jsx
@@ -1,12 +1,15 @@
 import React from 'react';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import { ApplicationTabs } from 'terra-application-links/lib/ApplicationLinks';
-import ApplicationLinkConfig from 'terra-application-links/lib/terra-dev-site/doc/common/ApplicationLinkConfig';
-import demoStyles from './demoStyles.scss';
+import ApplicationLinkConfig from 'terra-application-links/lib/terra-dev-site/doc/common/ApplicationLinksWithIconsConfig';
+import classNames from 'classnames/bind';
+import styles from './demoStyles.scss';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 
+const cx = classNames.bind(styles);
+
 const ApplicationTabsExample = () => (
-  <div className={demoStyles['demo-header']}>
+  <div className={cx(['demo-header', 'demo-header-with-icons'])}>
     <ApplicationTabs alignment="center" links={ApplicationLinkConfig} />
   </div>
 );