+++ packages/terra-application-links/tests/jest/testLinksWithIconsConfig.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import Icon from 'terra-icon/lib/icon/IconFeatured';
+
+const testLinksWithIconsConfig = [
+  {
+    path: '/test1',
+    text: 'test 1',
+    icon: <Icon />,
+  },
+  {
+    path: '/test2',
+    text: 'test 2',
+    icon: <Icon />,
+  },
+  {
+    path: '/test3',
+    text: 'test 3',
+    icon: <Icon />,
+  },
+  {
+    path: '/test4',
+    text: 'test 4',
+    icon: <Icon />,
+  },
+  {
+    path: '/test5',
+    text: 'test 5',
+    icon: <Icon />,
+  },
+  {
+    path: '/test6',
+    text: 'test 6',
+    icon: <Icon />,
+  },
+];
+
+export default testLinksWithIconsConfig;