+++ packages/terra-application-links/CHANGELOG.md
@@ -19,6 +19,9 @@ Unreleased
 * Update Jest tests to use terra-enzyme-intl
 * Changed demo mock background colors to be themable
 
+### Added
+* Support for icons in application links
+
 5.5.0 - (March 12, 2019)
 ------------------
 ### Changed